+++ .travis.sh
@@ -1,13 +1,35 @@
-#! /usr/bin/env sh
+#! /bin/zsh
 
-set -eo pipefail
-# set -euxo pipefail
+set -euo pipefail
 
-echo $TRAVIS_COMMIT_MESSAGE
-echo "RFCI_TASK = $RFCI_TASK"
-readonly RFWorkspace="RFKit.xcworkspace"
-readonly RFSTAGE="$1"
-echo "RFSTAGE = $RFSTAGE"
+logInfo () {
+    echo "\033[32m$1\033[0m" >&2
+}
+
+logWarning () {
+    echo "\033[33m$1\033[0m" >&2
+}
+
+logError () {
+    echo "\033[31m$1\033[0m" >&2
+}
+
+# Make sure all parameters are set correctly.
+logInfo "RFCI_PRODUCT_NAME = $RFCI_PRODUCT_NAME"
+
+readonly RFCI_TASK="${RFCI_TASK:? is not set.}"
+logInfo "RFCI_TASK = $RFCI_TASK"
+
+readonly RFCI_STAGE="${1:?STAGE is not set.}"
+logInfo "RFCI_STAGE = $RFCI_STAGE"
+
+readonly RFWorkspace=${RFWorkspace:="$RFCI_PRODUCT_NAME.xcworkspace"}
+logInfo "RFWorkspace = $RFWorkspace"
+
+TRAVIS_COMMIT_MESSAGE=${TRAVIS_COMMIT_MESSAGE:="$(logWarning 'TRAVIS_COMMIT_MESSAGE is not set, leave it blank.')"}
+TRAVIS_BRANCH=${TRAVIS_BRANCH:="$(logWarning 'TRAVIS_BRANCH is not set, leave it blank.')"}
+
+echo ""
 
 # Run test
 # $1 scheme
@@ -23,27 +45,51 @@ XC_TestMac() {
 
 # Run watchOS build
 XC_TestWatch() {
-    xcodebuild build -workspace "$RFWorkspace" -scheme Target-watchOS ONLY_ACTIVE_ARCH=NO | xcpretty
+    xcodebuild build -workspace "$RFWorkspace" -scheme "Target-watchOS" ONLY_ACTIVE_ARCH=NO | xcpretty
+}
+
+# Run tests on iOS Simulator.
+# The destinations are the first and last available destination that are automatically detected.
+# $1 scheme
+XC_TestAutoIOS() {
+    logInfo "Detecting destinations..."
+    destList=$(xcodebuild -showdestinations -workspace "$RFWorkspace" -scheme "$1" | grep "iOS Simulator")
+    destCount=$(echo "$destList" | wc -l)
+    destFirst=$(echo "$destList" | head -1)
+    destLast=$(echo "$destList" | tail -2 | head -1)
+    destFirstID=$(echo "$destFirst" | awk 'match($0,/id\:[0-9A-F-]+/){ print substr($0,RSTART+3,RLENGTH-3) }')
+    destLastID=$(echo "$destLast" | awk 'match($0,/id\:[0-9A-F-]+/){ print substr($0,RSTART+3,RLENGTH-3) }')
+
+    logWarning "Test on simulator (id: $destFirstID)."
+    XC_Test "$1" "platform=iOS Simulator,id=$destFirstID"
+
+    logWarning "Test on simulator (id: $destLastID)."
+    XC_Test "$1" "platform=iOS Simulator,id=$destLastID"
+}
+
+STAGE_SETUP() {
+    gem install cocoapods --no-document
 }
 
 STAGE_MAIN() {
     if [ "$RFCI_TASK" = "POD_LINT" ]; then
         if [[ "$TRAVIS_COMMIT_MESSAGE" = *"[skip lint]"* ]]; then
-            echo "Skip pod lint"
+            logWarning "Skip pod lint"
         else
-            echo "TRAVIS_BRANCH = $TRAVIS_BRANCH"
-            gem install cocoapods --no-document --quiet
-            if [ "$TRAVIS_BRANCH" = "develop" ]; then
-                pod lib lint --fail-fast --allow-warnings
-            else
+            if [[ "$TRAVIS_BRANCH" =~ ^[0-9]+\.[0-9]+ ]]; then
+                logWarning "Release the podspec."
+                pod trunk push "$RFCI_PRODUCT_NAME.podspec"
+            elif [ "$TRAVIS_BRANCH" = "master" ]; then
+                logInfo "Lint the podspec."
                 pod lib lint --fail-fast
+            else
+                logInfo "Lint the podspec."
+                pod lib lint --fail-fast --allow-warnings
             fi
         fi
 
     elif [ "$RFCI_TASK" = "Xcode11" ]; then
         pod install
-
-        echo "Test for macOS and watchOS."
         XC_TestMac
         XC_TestWatch
 
@@ -87,20 +133,13 @@ STAGE_MAIN() {
         XC_Test "Test-iOS"   "platform=iOS Simulator,name=iPhone X,OS=11.4"
         XC_Test "Test-tvOS"  "platform=tvOS Simulator,name=Apple TV 4K,OS=11.4"
 
-    elif [ "$RFCI_TASK" = "Xcode8" ]; then
-        pod install
-
-        echo "Test for macOS and watchOS."
-        XC_TestMac
-        XC_TestWatch
-
-        XC_Test "Test-iOS"   "platform=iOS Simulator,name=iPhone 6,OS=10.2"
-        XC_Test "Test-tvOS"  "platform=tvOS Simulator,name=Apple TV 1080p,OS=10.2"
+    else
+        logError "Unexpected CI task: $RFCI_TASK"
     fi
 }
 
 STAGE_SUCCESS() {
-    if [ "$RFCI_TASK" = "Xcode10" ]; then
+    if [ "${RFCI_COVERAGE-}" = "1" ]; then
         curl -s https://codecov.io/bash | bash -s
     fi
 }
@@ -111,4 +150,4 @@ STAGE_FAILURE() {
     fi
 }
 
-"STAGE_$RFSTAGE"
+"STAGE_$RFCI_STAGE"